requirements.txt
streamlit==1.46.1
streamlit-image-coordinates
Pillow
numpy
matplotlib